feature(automation): add automation for fetching issue patch files and creating a merge request.
this automation enables you to search all open issue for uploaded patch
files and if the users is a developer of higher, a Merge Request will be
created from the patch file.

to prevent the task from reusing already applied patch files, once the
mr is created the patch file is removed from the comment.

closes #19